Update display of params in UI based on parent param value
Some plugin parameters should be displayed or not displayed based on the
value of another parameter (for example, the value of the "algorithm"
parameter of AstraReconCpu can affect if the "centre_pad" parameter is
displayed or not).

Process list information in the UI is created from a python object
representation of a process list. When plugin parameter values in the UI
are updated, this could consequently modify the "display" attribute of
plugin parameters in the python object representation of the process
list, and so the UI should be updated accordingly.