Update process list filepath input field when folder icon is clicked
Note that the component "plEditorFilepathInput" is largely a copy of the
component "labelledInputFieldAndButton"; the need for the process list
filepath input field to be updateable by the folder icons in the LHS
pane as well as by inputting text into the input field made it difficult
to reuse labelledInputFieldAndButton. Ideally, both these components can
be merged through some form of generalisation sometime in the future.