Process list deletion refreshes search results
Note that the component "plTabSearchInput" is an almost exact copy of
the component "plEditorFilepathInput", the only differences are:
- the actions dispatched when the button is clicked
- the actions dispatched when the "input" event for the input field is
  fired
- the computed value that comes from the store

A generalisation would be preferred to merge these two very similar
components.

The actions can be passed as string parameters to a more general version
of the component by using props, so generalising them likely presents no
problem.

It is the computed value from the store that is more problematic: the
two components refer to two distinct variables in the store,
parameterisation of the computed values in some way to pass them as
props would potentially allow the generalisation to be done.