Generalise plTabSearchInput and plEditorFilepathInput
The component "labelledInputFieldAndButton" now replaces both of the
mentioned components, which were almost identical to each other.

In terms of the code changes to do the generalisaton, instead of making
labelledInputFieldAndButton handle:
- dispatching actions
- fetching the state of the input field text from the store

the chosen approach was to let a labelledInputFieldAndButton instance
emit the necessary events for the parent to then do these things,
similar to what is described in the following:
https://forum.vuejs.org/t/generic-reusable-data-bound-custom-components-with-vuex/1828/3.
This alleviated the issue of having a specific store state variable
needing to be a "computed" property in labelledInputFieldAndButton, and
instead it could be fetched by the parent and then passed down to the
labelledInputFieldAndButton instance via props.