Generalise labelledInputFieldAndButton to allow multiple buttons
Instead of having the buttons emit click events back up to the parent
component, have the parent pass a function (as part of a prop) for each
button, which can then be executed when the button is clicked.

This allows each button to have its own listener, rather than making all
the buttons emit the same custom event but then not being able to define
v-on for each individual button when putting labelledInputFieldAndButton
into the template of the parent component.