Highlight edited parameter value text
A few things worth noting:

Without the added ":key" attribute to the instances of
"pluginParamInputField", if a parameter value is edited and submitted
but not saved, when reloading the process list the same instance of the
"pluginParamInputField" component is reused. This means that the value
of "data.latestInputValue" is still the last value in the input field
from before reloading the process list, but the value of the "value"
prop is the last saved value in the process list's nexus file.

Since the process list wasn't saved before reloading it in the UI but a
parameter input field value was edited, these two values won't match.
This causes the text to be displayed in an edited state (italicised and
grey), despite the text value itself genuinely being the last saved
value of that parameter.

Adding in the key makes it such that the "pluginParamInputField"
instance is destroyed and recreated when there are unsaved edits to
parameter input fields and the process list is reloaded, so then
"data.latestInputValue" is reset back to the same value as the prop.
This causes the text value to be in an unedited state (no italicisation
and black) and still reflect the last saved value in the process list's
nexus file.

There's currently no noticeable performance issues from having the input
fields be destroyed and recreated more often than without the use of the
":key" attribute, but it is something worth keeping in mind if
performance issues begin to crop up with the parameter input fields.