Allow plugin entries in editor to be collapsible/expandable
Note that $refs has been used to implement the global collapse and
expand buttons so that the parent "plEditorTabContent" instance is able
to modify the "collapsed" data attribute in all the child
"plEditorPluginEntry" instances. For a first iteration of the
functionality it does the job fine, but there is likely a nicer way of
achieving the same functionality!