Use parameter type error strings from Savu
The following Savu commit
https://github.com/DiamondLightSource/Savu/commit/58dcbdb21e8e4154f76e21b05ae7698d568a19fa
exposes the string error messages that are created and displayed by the
configurator when a parameter type error occurs. It is used in this
change to display the same type error message in the UI that the
configurator would, rather than creating another type error string in
the UI code.