Add API function for adding plugin to a process list
Prior to this change, the API function api_savu.getPluginDetails() was
being used for both viewing a plugin's info in the plugins.html page and
for fetching a plugin's info to be added to the process list editor in
the processing.html page.

The plugin info displayed in the plugins.html page is a subset of the
plugin info displayed in the process list editor in the processing.html
page, so using the same API function for both causes an insufficient
amount of information to be given to the process list editor (such as
whether any of the plugin's parameters has options), so it makes sense
to separate these cases.