Add "Save As" button to file browser
The "Save As" button will always open the file browser, regardless of if
the input field outside the file browser has a filepath in it or not.
The "Save" button on the other hand allows the ability to quickly save a
process list without opening the file browser, provided that a filepath
is given.