Make adding a plugin count as an unsaved change
There is a prompt if a user has made modifications to a process list and
attempts to open another file without having saved those changes.
However, prior to this commit, adding a plugin to a process list didn't
get caught as a modification to a process list, and thus if a user added
a plugin to a process list and then attempted to open another file
without saving, that change would be silently lost.