Fix param descriptions not updating when reordered
Previously, if a parameter's value modification caused other parameters
to be shown or hidden via the "display" attribute on the server-side,
the parameter descriptions potentially wouldn't be associated with the
correct parameters anymore in their tooltips. Using a more specific :key
prop corrects the rendering issues and fixes this bug.