Pass stringified param val back to UI if type error occurs
When a parameter value modification is requested in the UI, part of the
process initiated in flask involves expecting a string containing the
new parameter value and then casting that to the expected type prior to
the check that the requested modification is a valid one.

Passing the cast value back to the UI when a type error occurs
eventually ends up causing non-string parameter values to be handed to
flask, which are then incorrectly assumed to be strings and then
attempted to be cast to the expected type, resulting in an error in the
Content.value() method.

Passing the string-version of the parameter value to the UI when a type
error occurs fixes this.