Convert param vals to str before cast when creating process lists from UI data
When using Content.value() to cast parameter values of type unicode to
the expected type for parameter values not using parameter tuning, it
works fine. However, if parameter values do use parameter tuning (ie,
it includes a semicolon character) and have type unicode,
Content.value() appears to always return a unicode value rather than a
str value, and so the cast parameter value is seen to be of an incorrect
type and thus the initialisation of that parameter value from the UI
data fails.

Converting parameter values to str before casting fixes this issue.