Remove unnecessary uses of str()
Now that the type of strings seen by flask are str and not unicode (like
what happened in python 2.7) it is unnecessary to convert parameter
value strings in process list data that came from the UI to str, prior
to casting using Content.value(). See commits fd0d4e3 and b801b54 for
when those uses of str() were originally added for python 2.7.