Split addPluginHelper() into two separate functions
The two parts are:
- createPlPluginElementsEntry(): the creation of the data structure
  format that is required of entries in state.plPluginElements
- addPluginToPlPluginElements(): the insertion of a data structure into
  state.plPluginElements

This was done because most changes to state.plPluginElements so far have
been focused on either adding or deleting plugin from a process list, so
it wasn't a problem to have these two operations lumped together.
However, for functionalities such as the "ref" command in the
configurator, a plugin's parameter info needs to be changed in place,
but the entire entry still needs to be recreated.

A change-in-place rather than a deletion-action+addition-action is
necessary because deleting the entire plugin entry in
state.plPluginElements and adding it back in with different parameter
info results in the plugin entry temporarily disappearing from the UI
before reappearing which isn't desirable.

Separately generating the data structure format that is needed for
entries in state.plPluginElements allows the choice between
adding/deleting entries, or resetting entries.

Adding entries can be done as before using addPluginHelper(), and
resetting entries can now be done by calling
createPlPluginElementsEntry() to get the right data structure format and
then using state.plPluginElements.splice() appropriately in a mutation.