Use watcher to update latestInputValue in PluginParamInputField
Previously, to get data.latestInputValue in PluginParamInputField to
update when a new value had been submitted, the PluginParamInputField
components had been given a key in PluginParamEditorTableRow such that
they would be destroyed and recreated upon every parameter value change.

This was in order to correctly identify when the value in the text input
field had been edited compared to the current submitted value in the
vuex store so then the text could be italicised and coloured grey to
notify users of the unsubmitted parameter value change.

(The reason that the input field's value wasn't simply set as the value
in the store was because it was desired to be able to edit the value
without necessarily submitting it, so a "local variable" of some sort
was required).

The value of latestInputValue for the newly created
PluginParamInputField would then receive the updated value via the
param.value prop, but still be editable and not necessarily submitting
to the store when changes had been made until the enter key had been
pressed or the input field lost focus.

However, a less inefficient way of achieving the same end result is to
have a watcher on the param.value prop and setting latestInputValue
based on if the prop has changed or not, avoiding the frequent
destruction and recreation of PluginParamInputField components via key
changes.