Fix bytes-related error when downloading citations
In the preparation for the Hebi developer session on 17/03/21 a new
hebi-api image was built and, due to there not being explicit package
versions in requirements.txt, some different package versions were
installed. At the same time, a new error relating to bytes appeared when
downloading citations in the UI, and it is likely that the two events
were related.

The changes in this commit fix that bytes-related error for the package
versions update in commit f7da874.