Add CAS authentication and session-activity checker to UI
Both CAS authentication and the session-activity checker are necessary
functionalities for deploying Hebi in a production environment, with the
former limiting access to a user's Hebi session, and the latter
periodically checking for inactive Hebi sessions that can be stopped in
order to free up cluster resources (if Hebi is deployed on a cluster).

The CAS-related changes in this commit are essentially making use of the
CAS authentication container added in the following commit:
https://github.com/yousefmoazzam/hebi/commit/0a8d76e97103a7c5ff4c08bbe9ea7108c3b443d7

The session-activity checker code in the UI works in conjunction with
the hebi-launcher application for deployment at Diamond:
https://github.com/yousefmoazzam/hebi-launcher/. Developers at other
facilities can make use of the session-activity checker in the UI by
emitting the "heartbeat-request" websocket event in whatever application
is created to monitor Hebi sessions at their facility.

The removal of leading backslashes in API requests (such as in
api_savu.js) is due to the leading backslashes causing the omission of
parts of the URL when deployed on Kubernetes and using an Ingress for
routing traffic (where each user is assigned a path in the Ingress
config). The leading backslashes removed the part of the URL that
identifies the user in the HTTP requests, causing the requests to not
arrive at their intended destination (within their own Ingress path).
Without the leading backslashes, the issue is resolved for running on
Kubernetes with an Ingress, and their omission doesn't affect running
Hebi locally.